[obj] Reduce chance of "successfully" loading non-OBJ files.
Fail if nothing of interest was loaded. Current ObjParser implementation makes this somewhat difficult, but this greatly reduces the chance of successfully returning an empty scene when a clearly not-an-OBJ file is fed to the parser.

Parsing will now only succeed if at least *one* of the following is true (we're still being pretty forgiving):

- At least one vertex was loaded, or
- At least one vertex normal was loaded, or
- At least one texture coord was loaded, or
- At least one face was loaded (index validity doesn't matter), or
- At least one mtllib referred to an existing file with length > 0 (material data doesn't matter).

This should reduce the vast majority of false positives without creating any false negatives.